requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
transformers==4.42.4
torch>=2.1.0
PyMuPDF==1.24.10
pdfminer.six==20231228
python-docx==1.1.2
celery==5.3.6
redis==5.0.4
uvicorn==0.30.1
django-cors-headers==4.3.1
